+++ packages/iov-bcp/types/connection.d.ts
@@ -15,7 +15,7 @@ export interface Account {
     readonly pubkey?: PublicKeyBundle;
     readonly balance: ReadonlyArray<Amount>;
 }
-export interface BcpTicker {
+export interface Token {
     readonly tokenTicker: TokenTicker;
     /**
      * A name to be displayed to the user which allows differentiation
@@ -172,8 +172,8 @@ export interface BlockchainConnection {
     readonly disconnect: () => void;
     readonly chainId: () => ChainId;
     readonly height: () => Promise<number>;
-    readonly getTicker: (ticker: TokenTicker) => Promise<BcpTicker | undefined>;
-    readonly getAllTickers: () => Promise<ReadonlyArray<BcpTicker>>;
+    readonly getTicker: (ticker: TokenTicker) => Promise<Token | undefined>;
+    readonly getAllTickers: () => Promise<ReadonlyArray<Token>>;
     /**
      * Get the current account information (e.g. balance)
      *